Modularity and Code Reusability Applied
